move `SessionValues` to submodule
